Update contributing-docs.md
Signed-off-by: tpmccallum tim.mccallum@fermyon.com

The documentation site now creates the in-house search index as part of the GitHub Action (`Deploy Developer Docs` -> `publish` -> `Create search index`).

The user no longer needs to run the `npm run build-index` on localhost before committing their contribution. It is still worth mentioning the search feature; which is a segway into how to increase visibility of content inside the in-house search.

Signed-off-by: Timothy McCallum <tim.mccallum@fermyon.com>