Add a check to see if a `spin` item is in the way of the installer succeeding (before we even download the spin tarball)
Signed-off-by: tpmccallum tim.mccallum@fermyon.com

I was finally able to reproduce an error that has been flagged in an issue. This update closes https://github.com/fermyon/developer/issues/748 by adding a new step in the installer, which checks to see if a `spin` item is in the way of the installer succeeding (before the user’s end expends any resources on downloading the tarball).

Signed-off-by: Timothy McCallum <tim.mccallum@fermyon.com>